add Eiger filename fix and Lambda2M definition
