+++ profile_bluesky/startup/instrument/devices/ad_lambda2M_delete.py
@@ -61,7 +61,7 @@ class TriggerDetectorState(TriggerBase):
 
     def setup_trigger(self):
         # Stage signals
-        self.cam.stage_sigs["trigger_mode"] = "Internal Series"
+        self.cam.stage_sigs["trigger_mode"] = "Internal"
         # self.cam.stage_sigs["manual_trigger"] = "Enable"
         # self.cam.stage_sigs["num_images"] = 1
         # self.cam.stage_sigs["num_exposures"] = 1
@@ -100,67 +100,48 @@ class TriggerDetectorState(TriggerBase):
         #     self._status.set_finished()
         #     self._status = None
 
-class AD_EpicsHdf5FileName_8IDI(AD_EpicsHdf5FileName):
-    def stage(self):
-        """
-        QZ: overrides the override in AD_EpicsHdf5FileName
-        that allows for writing file name with desired format
-        """
-        # Make a filename.
-        filename, read_path, write_path = self.make_filename()
 
-        # Ensure we do not have an old file open.
-        set_and_wait(self.capture, 0)
-        # These must be set before parent is staged (specifically
-        # before capture mode is turned on. They will not be reset
-        # on 'unstage' anyway.
-        set_and_wait(self.file_path, write_path)
-        set_and_wait(self.file_name, filename)
-        # set_and_wait(self.file_number, 0)
-
-        # get file number now since it is incremented during stage()
-        # file_number = self.file_number.get()
-        # Must avoid parent's stage() since it sets file_number to 0
-        # Want to call grandparent's stage()
-        # super().stage()     # avoid this - sets `file_number` to zero
-        # call grandparent.stage()
-        FileStoreBase.stage(self)
-
-        # AD does the file name templating in C
-        # We can't access that result until after acquisition
-        # so we apply the same template here in Python.
-        template = self.file_template.get()
-        self._fn = template % (read_path, filename)
-        self._fp = read_path
-        if not self.file_path_exists.get():
-            raise IOError(f"Path {self.file_path.get()} does not exist on IOC.")
-
-        self._point_counter = itertools.count()
-
-        # from FileStoreHDF5.stage()
-        res_kwargs = {"frame_per_point": self.get_frames_per_point()}
-        self._generate_resource(res_kwargs)   
-
-
-    def generate_datum(self, key, timestamp, datum_kwargs):
-        """Generate a uid and cache it with its key for later insertion."""
-        template = self.file_template.get()
-        filename, read_path, write_path = self.make_filename()
-        # file_number = self.file_number.get()
-        # hdf5_file_name = template % (read_path, filename, file_number)
-        hdf5_file_name = template % (read_path, filename)
-
-        # inject the actual name of the HDF5 file here into datum_kwargs
-        # datum_kwargs["HDF5_file_name"] = hdf5_file_name
-        datum_kwargs["filename"] = hdf5_file_name
-
-        logger.debug("make_filename: %s", hdf5_file_name)
-        logger.debug("write_path: %s", write_path)
-        return super(AD_EpicsHdf5FileName, self).generate_datum(key, timestamp, datum_kwargs)
-
-
-class myHdf5EpicsIterativeWriter(AD_EpicsHdf5FileName_8IDI, FileStoreIterativeWrite):
+class myHdf5EpicsIterativeWriter(AD_EpicsHdf5FileName, FileStoreIterativeWrite):
     pass
+    # def stage(self):
+    #     """
+    #     QZ: overrides the override in AD_EpicsHdf5FileName
+    #     that allows for writing file name with desired format
+    #     """
+    #     # Make a filename.
+    #     filename, read_path, write_path = self.make_filename()
+
+    #     # Ensure we do not have an old file open.
+    #     set_and_wait(self.capture, 0)
+    #     # These must be set before parent is staged (specifically
+    #     # before capture mode is turned on. They will not be reset
+    #     # on 'unstage' anyway.
+    #     set_and_wait(self.file_path, write_path)
+    #     set_and_wait(self.file_name, filename)
+    #     # set_and_wait(self.file_number, 0)
+
+    #     # get file number now since it is incremented during stage()
+    #     file_number = self.file_number.get()
+    #     # Must avoid parent's stage() since it sets file_number to 0
+    #     # Want to call grandparent's stage()
+    #     # super().stage()     # avoid this - sets `file_number` to zero
+    #     # call grandparent.stage()
+    #     FileStoreBase.stage(self)
+
+    #     # AD does the file name templating in C
+    #     # We can't access that result until after acquisition
+    #     # so we apply the same template here in Python.
+    #     template = self.file_template.get()
+    #     self._fn = template % (read_path, filename)
+    #     self._fp = read_path
+    #     if not self.file_path_exists.get():
+    #         raise IOError(f"Path {self.file_path.get()} does not exist on IOC.")
+
+    #     self._point_counter = itertools.count()
+
+    #     # from FileStoreHDF5.stage()
+    #     res_kwargs = {"frame_per_point": self.get_frames_per_point()}
+    #     self._generate_resource(res_kwargs)    
 
 
 class MyHDF5Plugin(HDF5Plugin_V34, myHdf5EpicsIterativeWriter):
@@ -206,7 +187,7 @@ class LocalEigerDetectorBase(DetectorBase):
 
         # # QZ added this to remove automatic FileNumber append 
         # # and to comply with DM transfer for Rigaku format
-        self.hdf1.stage_sigs["file_template"] = "%s%s.h5"
+        # self.hdf1.stage_sigs["file_template"] = "%s%s.h5"
 
         # This must always come last
         self.hdf1.stage_sigs["capture"]=self.hdf1.stage_sigs.pop('capture')
@@ -339,3 +320,10 @@ class LocalEigerDetectorBase(DetectorBase):
 
 class EigerDetector(TriggerDetectorState, LocalEigerDetectorBase):
     pass
+
+lambda2M = EigerDetector(
+    "8idLambda2m:",
+    write_path_template="",
+    read_path_template="",
+    name="lambda2M"
+)
